Final restructuring of files and folders. Added empty data files
