+++ courier/generate_cycles.py
@@ -1,6 +1,7 @@
 """
 
 """
+# TODO: SEEMS DEPRECATED
 
 
 def create_single_cycles(id, points_set, cross_docking_points, distances):
